move template questions/template sessions to separate JSON file
